+++ src/plot.rs
@@ -1,8 +1,9 @@
 use crate::circle;
 use crate::point;
+use crate::utils::FloatType;
 use plotters::prelude::*;
 
-pub fn draw_plot(main_circle_radius: f64, circles: &Vec<circle::Circle>) {
+pub fn draw_plot(main_circle_radius: FloatType, circles: &Vec<circle::Circle>) {
     let plot_size: i32 = 1000;
     let root =
         BitMapBackend::new("circle.png", (plot_size as u32, plot_size as u32)).into_drawing_area();
@@ -25,8 +26,8 @@ pub fn draw_plot(main_circle_radius: f64, circles: &Vec<circle::Circle>) {
 
     let root = chart.plotting_area();
 
-    let convert_radius = |radius: f64| {
-        return (radius * (plot_size - 2 * 85) as f64) / (2.0 * spec_size);
+    let convert_radius = |radius: FloatType| {
+        return (radius * (plot_size - 2 * 85) as FloatType) / (2.0 * spec_size);
     };
 
     let set_circle = |c: &circle::Circle| {
