+++ src/ralgo/ralgo_params.rs
@@ -1,9 +1,11 @@
+use crate::utils::FloatType;
+
 #[derive(Debug)]
 pub struct RalgoParams {
-    pub alpha: f64,
-    pub q1: f64,
-    pub epsx: f64,
-    pub epsg: f64,
+    pub alpha: FloatType,
+    pub q1: FloatType,
+    pub epsx: FloatType,
+    pub epsg: FloatType,
     pub max_iterations: usize,
 }
 
@@ -20,11 +22,11 @@ impl Default for RalgoParams {
 }
 
 impl RalgoParams {
-    pub fn with_alpha(&self, alpha: f64) -> Self {
+    pub fn with_alpha(&self, alpha: FloatType) -> Self {
         RalgoParams { alpha, ..*self }
     }
 
-    pub fn with_q1(&self, q1: f64) -> Self {
+    pub fn with_q1(&self, q1: FloatType) -> Self {
         RalgoParams { q1, ..*self }
     }
 
@@ -35,14 +37,14 @@ impl RalgoParams {
         }
     }
 
-    pub fn with_epsx(&self, epsx: f64) -> Self {
+    pub fn with_epsx(&self, epsx: FloatType) -> Self {
         RalgoParams {
             epsx,
             ..*self
         }
     }
 
-    pub fn with_epsg(&self, epsg: f64) -> Self {
+    pub fn with_epsg(&self, epsg: FloatType) -> Self {
         RalgoParams {
             epsg,
             ..*self
